minor bug fixes

+++ MPU9250.cpp
@@ -473,10 +473,12 @@ void MPU9250::calibrateAK8963Mag(float * dest1, float * dest2)
 {
 	uint16_t ii = 0, sample_count = 0;
 	int32_t mag_bias[3] = {0, 0, 0}, mag_scale[3] = {0, 0, 0};
-	short mag_max = new short[3];
+
+	int mag_max = new int[3];
 	mag_max[0] = -32767; mag_max[1] = -32767; mag_max[2] = -32767;
-	short mag_min  = new short[3];
+	int mag_min  = new int[3];
 	mag_min[0] = 32768; mag_min[1] = 32768; mag_min[2] = 32768;
+	
 	int16_t mag_temp[3] = {0, 0, 0};
 
 	std::cout << "Mag Calibration: Wave device in a figure eight until done!" << std::endl;
